+++ Network/Telegram/API/Bot/Property/Identifiable.hs
@@ -5,7 +5,7 @@ import "base" Data.Int (Int, Int64)
 import Network.Telegram.API.Bot.Object (Object)
 import Network.Telegram.API.Bot.Object.Chat (Chat (Private, Group, Supergroup, Channel))
 import Network.Telegram.API.Bot.Object.From (From (Bot, User))
-import Network.Telegram.API.Bot.Object.Message (Message (Direct))
+import Network.Telegram.API.Bot.Object.Message (Message (Direct, Forward))
 import Network.Telegram.API.Bot.Object.Update (Update (Query, Membership, Incoming))
 
 type family Identificator o = i
@@ -31,6 +31,7 @@ instance Identifiable From where
 
 instance Identifiable Message where
 	identificator (Direct i _ _ _) = i
+	identificator (Forward i _ _ _) = i
 
 instance Identifiable Update where
 	identificator (Query i _) = i